fix(interactive-tour-edit): memoize tour step edit controls
